Fixed warnings in the iFP target.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@8671 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/pcm_playback.c
@@ -562,8 +562,8 @@ void pcm_set_frequency(unsigned int frequency)
     (void)frequency;
 }
 
-void pcm_play_data(void (*get_more)(unsigned char** start, long* size),
-        unsigned char* start, long size)
+void pcm_play_data(void (*get_more)(unsigned char** start, size_t* size),
+        unsigned char* start, size_t size)
 {
     (void)get_more;
     (void)start;
@@ -600,7 +600,7 @@ void pcm_calculate_peaks(int *left, int *right)
     (void)right;
 }
 
-long pcm_get_bytes_waiting(void)
+size_t pcm_get_bytes_waiting(void)
 {
     return 0;
 }
